Neopixel LEDs
Added a strip of neopixels for decoration. Needed to include the Adafruit_neopixel library along with the esp.c library that it uses. In the config, I had to set CONFIG_PM_ENABLE=n because the power managment thats enabled with Wifi/BLE affects the RMT that the neopixel uses for 800kHz timer.