Watchdog Fix and Adjusted Battery Reporting
I modified the watchdog to reset the ESP32 properly if there is a block. This was tested and confirmed. The settings were move to the sdkconfig file.
A warning battery state was included to provide notice of when the battery is getting low.
Modified the ESP LOG reporting so it now prints out as specified in the main sketch.